tectonic/core-memory.c: rename from core-kpathsea.c
The last bit in "internals.h" was some dynamic memory routines. Break those
out into their own files. Now "internals.h" is basicaly superfluous.
