+++ tectonic/internals.h
@@ -8,27 +8,12 @@
 
 #include "tectonic.h"
 #include "core-bridge.h"
+#include "core-memory.h"
 #include "core-strutils.h"
 
 #include <fcntl.h>
 #include <stdio.h>
 #include <sys/stat.h>
 #include <sys/types.h>
-
-
-BEGIN_EXTERN_C
-
-/* core-kpathutil.c */
-char *xstrdup (const char *s);
-void *xmalloc (size_t size);
-void *xrealloc (void *old_address, size_t new_size);
-void *xcalloc (size_t nelem, size_t elsize);
-
-static inline void *mfree(void *ptr) {
-    free(ptr);
-    return NULL;
-}
-
-END_EXTERN_C
 
 #endif /* not TECTONIC_INTERNALS_H */
